feat(function-entry): added hints popover
